[cleanup] Improve the organization of the code of the tree item context menu
