+++ packages/trees/frontend/sirius-components-trees/src/index.ts
@@ -15,13 +15,14 @@ export * from './toolbar/TreeToolBarContext';
 export * from './toolbar/TreeToolBarContext.types';
 export * from './toolbar/TreeToolBarContribution';
 export * from './toolbar/TreeToolBarContribution.types';
+export * from './treeitems/context-menu/TreeItemContextMenu';
+export * from './treeitems/context-menu/TreeItemContextMenu.types';
+export * from './treeitems/context-menu/TreeItemContextMenuEntry.types';
+export * from './treeitems/context-menu/TreeItemContextMenuEntryExtensionPoints';
+export * from './treeitems/context-menu/TreeItemContextMenuEntryExtensionPoints.types';
+export { type GQLTreeItemContextMenuEntry } from './treeitems/context-menu/useContextMenuEntries.types';
 export * from './treeitems/filterTreeItem';
 export * from './treeitems/TreeItemAction.types';
-export * from './treeitems/TreeItemContextMenu';
-export * from './treeitems/TreeItemContextMenu.types';
-export * from './treeitems/TreeItemContextMenuEntry.types';
-export * from './treeitems/TreeItemContextMenuEntryExtensionPoints';
-export * from './treeitems/TreeItemContextMenuEntryExtensionPoints.types';
 export * from './trees/FilterBar';
 export * from './trees/TreeRepresentation';
 export * from './views/getTreeEventSubscription';
@@ -30,8 +31,8 @@ export * from './views/TreeFiltersMenu.types';
 export * from './views/TreeView';
 export * from './views/TreeView.types';
 export * from './views/TreeViewExtensionPoints';
-export * from './views/useTreePath';
-export * from './views/useTreePath.types';
 export * from './views/useExpandAllTreePath';
 export * from './views/useExpandAllTreePath.types';
 export * from './views/useTreeFilters';
+export * from './views/useTreePath';
+export * from './views/useTreePath.types';
