+++ packages/trees/frontend/sirius-components-trees/src/treeitems/context-menu/DefaultMenuItem.types.ts
@@ -1,0 +1,24 @@
+/*******************************************************************************
+ * Copyright (c) 2025 Obeo.
+ * This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v2.0
+ * which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     Obeo - initial API and implementation
+ *******************************************************************************/
+
+import { GQLTreeItem } from '../../views/TreeView.types';
+import { GQLTreeItemContextMenuEntry } from './useContextMenuEntries.types';
+
+export interface DefaultMenuItemProps {
+  editingContextId: string;
+  treeId: string;
+  item: GQLTreeItem;
+  entry: GQLTreeItemContextMenuEntry;
+  readOnly: boolean;
+  onClick: () => void;
+}
