+++ packages/trees/frontend/sirius-components-trees/src/treeitems/context-menu/RenameMenuItem.tsx
@@ -1,0 +1,31 @@
+/*******************************************************************************
+ * Copyright (c) 2025 Obeo.
+ * This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v2.0
+ * which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     Obeo - initial API and implementation
+ *******************************************************************************/
+import EditIcon from '@mui/icons-material/Edit';
+import ListItemIcon from '@mui/material/ListItemIcon';
+import ListItemText from '@mui/material/ListItemText';
+import MenuItem from '@mui/material/MenuItem';
+import { RenameMenuItemProps } from './RenameMenuItem.types';
+
+export const RenameMenuItem = ({ item, readOnly, onClick }: RenameMenuItemProps) => {
+  if (!item.editable) {
+    return null;
+  }
+  return (
+    <MenuItem key="rename" onClick={onClick} data-testid="rename-tree-item" disabled={readOnly} aria-disabled>
+      <ListItemIcon>
+        <EditIcon fontSize="small" />
+      </ListItemIcon>
+      <ListItemText primary="Rename" />
+    </MenuItem>
+  );
+};
