+++ packages/trees/frontend/sirius-components-trees/src/treeitems/context-menu/TreeItemContextMenuEntryExtensionPoints.ts
