+++ packages/trees/frontend/sirius-components-trees/src/treeitems/context-menu/useContextMenuEntries.types.ts
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2024 Obeo.
+ * Copyright (c) 2024, 2025 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -10,10 +10,9 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-import { TreeItemContextMenuEntry } from './TreeItemContextMenu.types';
 
 export interface UseContextMenuEntriesValue {
-  contextMenuEntries: TreeItemContextMenuEntry[];
+  contextMenuEntries: GQLTreeItemContextMenuEntry[];
   loading: boolean;
 }
